brick-loader/src/main/java/de/csgis/geobricks/servlet/IndexHTMLRedirectFilter.java
package de.csgis.geobricks.servlet;

import java.io.IOException;

import javax.inject.Singleton;
import javax.servlet.Filter;
import javax.servlet.FilterChain;
import javax.servlet.FilterConfig;
import javax.servlet.ServletException;
import javax.servlet.ServletRequest;
import javax.servlet.ServletResponse;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

/**
 * Filter for redirecting requests without an ending slash (<code>/</code>) to
 * the same URL with an ending slash.
 * 
 * @author vicgonco
 */
@Singleton
public class IndexHTMLRedirectFilter implements Filter {

	@Override
	public void init(FilterConfig filterConfig) throws ServletException {
		// do nothing
	}

	@Override
	public void doFilter(ServletRequest request, ServletResponse response,
			FilterChain chain) throws IOException, ServletException {
		String uri = ((HttpServletRequest) request).getRequestURI();
		if (uri.endsWith("/")) {
			chain.doFilter(request, response);
		} else {
			HttpServletResponse resp = (HttpServletResponse) response;
			resp.sendRedirect(uri + "/");
		}
	}

	@Override
	public void destroy() {
		// do nothing
	}
}
